quantum-attendance/backend/requirements.txt
flask==2.3.3
qiskit==0.45.0
qiskit-ibm-runtime==0.15.0
python-dotenv==1.0.0
APScheduler==3.10.4
gspread==5.12.2
oauth2client==4.1.3
Flask-Cors==4.0.0
gunicorn==21.2.0
